+++ src/main/java/net/sashiro/radioactiveores/events/ModEvents.java
@@ -11,9 +11,7 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.attributes.AttributeInstance;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraftforge.event.entity.EntityAttributeModificationEvent;
@@ -25,11 +23,43 @@ import net.sashiro.radioactiveores.attributes.Attributes;
 import net.sashiro.radioactiveores.block.Radioactive;
 import net.sashiro.radioactiveores.block.RadioactiveBlock;
 import net.sashiro.radioactiveores.effects.MobEffects;
+import net.sashiro.radioactiveores.item.RadioactiveBlockItem;
+import net.sashiro.radioactiveores.item.RadioactiveItem;
+import net.sashiro.radioactiveores.util.Config;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.ArrayList;
 
 @SuppressWarnings({"rawtypes", "unchecked", "DataFlowIssue"})
 @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD, modid = RadioactiveOres.MOD_ID)
 public class ModEvents {
     private static int tick = 0;
+    private static int tick2 = 0;
+
+    @NotNull
+    private static MobEffectInstance addRadiation(int amplifier) {
+        return new MobEffectInstance(MobEffects.RADIATION_EFFECT.get(), 6000, amplifier);
+    }
+
+    private static void applyRadiationEffect(LivingEntity entity, double radiationLevel, boolean hasRadiationEffect, int amplifier) {
+        if (entity.hasEffect(MobEffects.RADIATION_BLOCKER_EFFECT.get())) {
+            if (hasRadiationEffect) entity.removeEffect(MobEffects.RADIATION_EFFECT.get());
+            return;
+        }
+        if (hasRadiationEffect) {
+            if (amplifier == 0 && radiationLevel >= 300) {
+                entity.removeEffect(MobEffects.RADIATION_EFFECT.get());
+                entity.addEffect(addRadiation(1));
+            } else if (amplifier == 1 && radiationLevel >= 600) {
+                entity.removeEffect(MobEffects.RADIATION_EFFECT.get());
+                entity.addEffect(addRadiation(2));
+            }
+        } else {
+            if (radiationLevel >= 100 && radiationLevel < 300) entity.addEffect(addRadiation(0));
+            if (radiationLevel >= 300 && radiationLevel < 600) entity.addEffect(addRadiation(1));
+            if (radiationLevel >= 600) entity.addEffect(addRadiation(2));
+        }
+    }
 
     @SubscribeEvent
     public void modifyAttributes(EntityAttributeModificationEvent event) {
@@ -40,68 +70,69 @@ public class ModEvents {
     }
 
     @SubscribeEvent
-    public void onLivingUpdate(LivingEvent.LivingTickEvent event) {
+    public void onLivingTick(LivingEvent.LivingTickEvent event) {
         LivingEntity entity = event.getEntity();
-        if (entity.getAttributes().hasAttribute(Attributes.RADIATION_ATTRIBUTE.get()) && !entity.hasEffect(MobEffects.RADIATION_BLOCKER_EFFECT.get()) && !entity.isInvulnerable()) {
-            if (entity instanceof Player player && player.isCreative()) return;
-            AttributeInstance instance = entity.getAttribute(Attributes.RADIATION_ATTRIBUTE.get());
-            double radiationLevel = instance.getValue();
-            boolean hasRadiationEffect = entity.hasEffect(MobEffects.RADIATION_EFFECT.get());
-            int amplifier = -1;
-            if (entity.hasEffect(MobEffects.RADIATION_EFFECT.get()))
-                amplifier = entity.getEffect(MobEffects.RADIATION_EFFECT.get()).getAmplifier();
-            if (tick == 20 || tick == 40) {
-                if ((hasRadiationEffect || radiationLevel >= 100) && radiationLevel < 300) {
-                    if (!hasRadiationEffect)
-                        entity.addEffect(new MobEffectInstance(MobEffects.RADIATION_EFFECT.get(), 6000, 0));
-                } else if (radiationLevel >= 300 && radiationLevel < 600) {
-                    if (!hasRadiationEffect) {
-                        entity.addEffect(new MobEffectInstance(MobEffects.RADIATION_EFFECT.get(), 6000, 1));
-                    } else if (amplifier == 0) {
-                        entity.removeEffect(MobEffects.RADIATION_EFFECT.get());
-                        entity.addEffect(new MobEffectInstance(MobEffects.RADIATION_EFFECT.get(), 6000, 1));
-                    }
-                } else if (radiationLevel >= 600) {
-                    if (!hasRadiationEffect) {
-                        entity.addEffect(new MobEffectInstance(MobEffects.RADIATION_EFFECT.get(), 6000, 2));
-                    } else if (amplifier == 0 || amplifier == 1) {
-                        entity.removeEffect(MobEffects.RADIATION_EFFECT.get());
-                        entity.addEffect(new MobEffectInstance(MobEffects.RADIATION_EFFECT.get(), 6000, 2));
-                    }
-                }
 
-                if (entity instanceof Player player) {
-                    if (getInventoryItem(player) instanceof Radioactive) {
-                        radiationLevel = radiationLevel + 1;
-                        instance.setBaseValue(radiationLevel);
+        if (!entity.getAttributes().hasAttribute(Attributes.RADIATION_ATTRIBUTE.get())) return;
+        if (entity instanceof Player player && player.isCreative()) return;
+        if (entity.isInvulnerable()) return;
 
-                    }
+        AttributeInstance instance = entity.getAttribute(Attributes.RADIATION_ATTRIBUTE.get());
+        int amplifier = -1;
+        double radiationLevel = instance.getValue();
+        boolean hasRadiationEffect = entity.hasEffect(MobEffects.RADIATION_EFFECT.get());
+
+        if (hasRadiationEffect) amplifier = entity.getEffect(MobEffects.RADIATION_EFFECT.get()).getAmplifier();
+
+        applyRadiationEffect(entity, radiationLevel, hasRadiationEffect, amplifier);
+
+        if (tick2 >= 20) {
+            if (entity instanceof Player player) {
+                ArrayList<RadioactiveItem> radioactiveItems = getRadioactiveItems(player);
+                ArrayList<RadioactiveBlockItem> radioactiveBlockItems = getRadioactiveBlockItems(player);
+                for (RadioactiveItem item : radioactiveItems) {
+                    int count;
+                    if (radioactiveItems.size() == 0) return;
+                    ItemStack stack = new ItemStack(item);
+                    count = stack.getCount();
+                    radiationLevel = radiationLevel + (count * item.getAmplifier());
                 }
+                for (RadioactiveBlockItem item : radioactiveBlockItems) {
+                    int count;
+                    if (radioactiveBlockItems.size() == 0) return;
+                    ItemStack stack = new ItemStack(item);
+                    count = stack.getCount();
+                    radiationLevel = radiationLevel + (count * item.getAmplifier());
+                }
+                instance.setBaseValue(radiationLevel);
+
             }
-            if (tick == 50) {
-                Block mcBlock;
-                RadioactiveBlock block;
+        }
 
-                if (entity instanceof Player) {
-                    mcBlock = testFor(entity, 3);
-                } else {
-                    mcBlock = testFor(entity, 2);
-                }
-                if (mcBlock instanceof Radioactive) {
-                    block = (RadioactiveBlock) mcBlock;
-                    radiationLevel = radiationLevel + block.getAmplifier();
-                    instance.setBaseValue(radiationLevel);
-                }
+        if (tick >= Config.ticksToCheckForRadiation) {
+            Block mcBlock;
+            RadioactiveBlock block;
+
+            if (entity instanceof Player) {
+                mcBlock = testForBlock(entity, Config.playerRadius);
+            } else {
+                mcBlock = testForBlock(entity, Config.entityRadius);
             }
-            if (tick == 50) tick = 0;
-            else tick++;
+            if (mcBlock instanceof Radioactive) {
+                block = (RadioactiveBlock) mcBlock;
+                radiationLevel = radiationLevel + block.getAmplifier();
+                instance.setBaseValue(radiationLevel);
+            }
         }
-        if (entity.hasEffect(MobEffects.RADIATION_EFFECT.get()) && entity.hasEffect(MobEffects.RADIATION_BLOCKER_EFFECT.get())) {
-            entity.removeEffect(MobEffects.RADIATION_EFFECT.get());
-        }
+
+        if (tick >= Config.ticksToCheckForRadiation + 1) tick = 0;
+        else tick++;
+
+        if (tick2 >= 21) tick2 = 0;
+        else tick2++;
     }
 
-    private Block testFor(LivingEntity entity, int distance) {
+    private Block testForBlock(LivingEntity entity, int distance) {
         int x = (int) entity.getX();
         int y = (int) entity.getY();
         int z = (int) entity.getZ();
@@ -119,13 +150,27 @@ public class ModEvents {
         return Blocks.AIR;
     }
 
-    private Item getInventoryItem(Player player) {
+    private ArrayList<RadioactiveItem> getRadioactiveItems(Player player) {
+        ArrayList<RadioactiveItem> radioactiveItems = new ArrayList<>();
         NonNullList<ItemStack> items = player.getInventory().items;
-        for (ItemStack i : items) {
-            if (i.getItem() instanceof Radioactive) {
-                return i.getItem();
+        for (ItemStack inventoryItem : items) {
+            if (inventoryItem.getItem() instanceof Radioactive) {
+                if (inventoryItem.getItem() instanceof RadioactiveItem)
+                    radioactiveItems.add((RadioactiveItem) inventoryItem.getItem());
             }
         }
-        return Items.AIR;
+        return radioactiveItems;
+    }
+
+    private ArrayList<RadioactiveBlockItem> getRadioactiveBlockItems(Player player) {
+        ArrayList<RadioactiveBlockItem> radioactiveItems = new ArrayList<>();
+        NonNullList<ItemStack> items = player.getInventory().items;
+        for (ItemStack inventoryItem : items) {
+            if (inventoryItem.getItem() instanceof Radioactive) {
+                if (inventoryItem.getItem() instanceof RadioactiveBlockItem)
+                    radioactiveItems.add((RadioactiveBlockItem) inventoryItem.getItem());
+            }
+        }
+        return radioactiveItems;
     }
 }