+++ changelog.md
@@ -1,1 +1,12 @@
-update to 1.19.4
+update: 1.0.2
+
+# Changes
+- code cleanups
+- added new configurable options
+    * **playerRadius** *(radius check for Player)*
+    * **entityRadius** *(radius check for Entity)*
+    * **ticksToCheckForRadiation** *(time in ticks to execute checks)*
+- Item count & amplifier in Inventory now affects the radiation level
+
+# Todo
+- add configurable radiation levels